sorting bar is done
