+++ src/interfaces/card-items/CardItems.ts
@@ -1,8 +1,11 @@
+type Stack = "js" | "ts" | "react" | "";
+
 export interface CardItems {
   id: number;
   src: string;
   text: string;
   lang: string;
   href: string;
+  stack: Stack;
   favorite?: string;
 }
