+++ src/pages/Favorite-page/FavoritePage.module.scss
@@ -49,6 +49,7 @@
   box-sizing: border-box;
   width: 30px;
 }
+
 .card_body__footer {
   display: flex;
   align-items: center;
