--- src/pages/Library-page/LibraryPage.module.scss
@@ -1,7 +1,0 @@
-.library_page {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  row-gap: 30px;
-  flex-wrap: wrap;
-}
