+++ src/pages/Library-page/LibraryPage.tsx
@@ -1,10 +1,10 @@
-import CourseCard from "../../ui/course-card/CourseCard";
+import CourseCard from "../../modules/course-card/CourseCard";
 import s from "./LibraryPage.module.scss";
 
 const LibraryPage = () => {
   return (
     <div className="_container pt36">
-      <div className={s.library_page}>
+      <div className={s.card_container}>
         <CourseCard type="library" />
       </div>
     </div>
